+++ filen-cli/src/docs.rs
@@ -96,6 +96,15 @@ static PARSED_DOC_OUTLINE: LazyLock<Result<Vec<ParsedDocSection>>> = LazyLock::n
 			],
 		},
 		DocSection {
+			id: "transfer-files",
+			title: "Transferring Files",
+			elements: vec![
+				DocElement::DocFragment("transfers"),
+				DocElement::CommandHelp("upload"),
+				DocElement::CommandHelp("download"),
+			],
+		},
+		DocSection {
 			id: "managed-rclone",
 			title: "Managed Rclone",
 			elements: vec![